Move shared test utils to utils_test.go
... because they will be shared between the
Service and StatefulSet tests.

Only moves the code with no changes, should not change
(test) behavior.
